Basic working version

+++ src/main/java/com/chuyachia/chip8emulator/Stack.java
@@ -3,14 +3,14 @@ package com.chuyachia.chip8emulator;
 public class Stack {
     private final byte STACK_LEVEL = 16;
     private byte pointer;
-    private final int[] stack;
+    private final short[] stack;
 
     public Stack() {
         this.pointer = 0;
-        this.stack = new int[STACK_LEVEL];
+        this.stack = new short[STACK_LEVEL];
     }
 
-    public int pop() throws Exception {
+    public short pop() throws Exception {
         if (pointer == 0) {
             throw new Exception("Stack is empty");
         }
@@ -18,7 +18,7 @@ public class Stack {
         return stack[pointer--];
     }
 
-    public void push(int value) throws Exception{
+    public void push(short value) throws Exception{
         if (pointer >= STACK_LEVEL-1) {
             throw new Exception("Stack overflow");
         }
